refactored changes into seperate file SSCS-10293
