+++ src/main/java/uk/gov/hmcts/reform/sscs/helper/mapping/HearingsDetailsMapping.java
@@ -4,15 +4,12 @@ import lombok.extern.slf4j.Slf4j;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.CcdValue;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Entity;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Hearing;
 import uk.gov.hmcts.reform.sscs.ccd.domain.OtherParty;
 import uk.gov.hmcts.reform.sscs.ccd.domain.OverrideFields;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Party;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.exception.InvalidMappingException;
-import uk.gov.hmcts.reform.sscs.model.HearingLocation;
 import uk.gov.hmcts.reform.sscs.model.HearingWrapper;
-import uk.gov.hmcts.reform.sscs.model.VenueDetails;
 import uk.gov.hmcts.reform.sscs.model.hmc.reference.HearingType;
 import uk.gov.hmcts.reform.sscs.model.single.hearing.HearingDetails;
 import uk.gov.hmcts.reform.sscs.service.holder.ReferenceDataServiceHolder;
@@ -20,7 +17,6 @@ import uk.gov.hmcts.reform.sscs.service.holder.ReferenceDataServiceHolder;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
 
@@ -29,8 +25,6 @@ import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.isYes;
 import static uk.gov.hmcts.reform.sscs.model.hmc.reference.HearingType.SUBSTANTIVE;
-import static uk.gov.hmcts.reform.sscs.model.hmc.reference.LocationType.COURT;
-import static uk.gov.hmcts.reform.sscs.model.hmc.reference.NextHearingVenueType.SOMEWHERE_ELSE;
 import static uk.gov.hmcts.reform.sscs.reference.data.model.HearingPriority.STANDARD;
 import static uk.gov.hmcts.reform.sscs.reference.data.model.HearingPriority.URGENT;
 
@@ -59,7 +53,7 @@ public final class HearingsDetailsMapping {
             .hearingPriorityType(getHearingPriority(caseData))
             .numberOfPhysicalAttendees(HearingsNumberAttendeesMapping.getNumberOfPhysicalAttendees(caseData))
             .hearingInWelshFlag(shouldBeHearingsInWelshFlag())
-            .hearingLocations(getHearingLocations(caseData, referenceDataServiceHolder))
+            .hearingLocations(HearingsLocationMapping.getHearingLocations(caseData, referenceDataServiceHolder))
             .facilitiesRequired(getFacilitiesRequired())
             .listingComments(getListingComments(caseData))
             .hearingRequester(getHearingRequester())
@@ -96,89 +90,6 @@ public final class HearingsDetailsMapping {
     public static boolean shouldBeHearingsInWelshFlag() {
         // TODO Future Work
         return false;
-    }
-
-    public static List<HearingLocation> getHearingLocations(SscsCaseData caseData,
-                                                            ReferenceDataServiceHolder referenceDataServiceHolder)
-                                                            throws InvalidMappingException {
-        String nextHearingVenueName = caseData.getAdjournCaseNextHearingVenue();
-
-        //TODO: remove flag
-        if (referenceDataServiceHolder.isAdjournmentFlagEnabled() && isNotEmpty(nextHearingVenueName)
-            && !HearingsChannelMapping.isPaperCase(caseData)) {
-            String epimsID = getEpimsID(caseData, nextHearingVenueName);
-
-            log.info("Getting hearing location with the venue ID of {}", epimsID);
-
-            return List.of(HearingLocation.builder()
-                .locationId(epimsID)
-                .locationType(COURT)
-                .build());
-        }
-
-        return getAllHearingLocations(caseData, referenceDataServiceHolder);
-    }
-
-    public static List<HearingLocation> getAllHearingLocations(SscsCaseData caseData,
-                                                                ReferenceDataServiceHolder referenceDataServiceHolder) {
-        OverrideFields overrideFields = OverridesMapping.getOverrideFields(caseData);
-
-        if (isNotEmpty(overrideFields.getHearingVenueEpimsIds())) {
-            return overrideFields.getHearingVenueEpimsIds().stream()
-                .map(CcdValue::getValue)
-                .map(CcdValue::getValue)
-                .map(epimsId -> HearingLocation.builder()
-                    .locationId(epimsId)
-                    .locationType(COURT)
-                    .build())
-                .collect(Collectors.toList());
-        }
-
-        if (HearingsChannelMapping.isPaperCase(caseData)) {
-            List<VenueDetails> venueDetailsList = referenceDataServiceHolder
-                .getVenueService()
-                .getActiveRegionalEpimsIdsForRpc(caseData.getRegionalProcessingCenter().getEpimsId());
-
-            log.info("Found {} venues under RPC {} for paper case {}", venueDetailsList.size(),
-                caseData.getRegionalProcessingCenter().getName(), caseData.getCcdCaseId());
-
-            return venueDetailsList.stream()
-                .map(VenueDetails::getEpimsId)
-                .map(id -> HearingLocation.builder()
-                    .locationId(id)
-                    .locationType(COURT)
-                    .build())
-                .collect(Collectors.toList());
-        }
-
-        String epimsId = referenceDataServiceHolder
-            .getVenueService()
-            .getEpimsIdForVenue(caseData.getProcessingVenue())
-            .orElse(null);
-
-        Map<String,List<String>> multipleHearingLocations = referenceDataServiceHolder.getMultipleHearingLocations();
-
-        return multipleHearingLocations.values().stream()
-            .filter(listValues ->  listValues.contains(epimsId))
-            .findFirst()
-            .orElseGet(() -> Collections.singletonList(epimsId))
-            .stream().map(epims -> HearingLocation.builder().locationId(epims).locationType(COURT).build())
-            .collect(Collectors.toCollection(ArrayList::new));
-    }
-
-    private static String getEpimsID(SscsCaseData caseData, String nextHearingVenue) throws InvalidMappingException {
-        if (SOMEWHERE_ELSE.getValue().equals(nextHearingVenue)) {
-            return caseData.getAdjournCaseNextHearingVenueSelected().getValue().getCode();
-        }
-
-        Hearing latestHearing = caseData.getLatestHearing();
-
-        if (nonNull(latestHearing)) {
-            return latestHearing.getValue().getEpimsId();
-        }
-
-        throw new InvalidMappingException("Failed to determine next hearing venue due to no latest hearing on case "
-                                          + caseData.getCcdCaseId());
     }
 
     public static List<String> getFacilitiesRequired() {
