+++ src/main/java/uk/gov/hmcts/reform/sscs/helper/mapping/ServiceHearingValuesMapping.java
@@ -44,7 +44,7 @@ public final class ServiceHearingValuesMapping {
                 .hearingPriorityType(HearingsDetailsMapping.getHearingPriority(caseData))
                 .numberOfPhysicalAttendees(HearingsNumberAttendeesMapping.getNumberOfPhysicalAttendees(caseData))
                 .hearingInWelshFlag(HearingsDetailsMapping.shouldBeHearingsInWelshFlag())
-                .hearingLocations(HearingsDetailsMapping.getAllHearingLocations(caseData, referenceDataServiceHolder))
+                .hearingLocations(HearingsLocationMapping.getHearingLocations(caseData, referenceDataServiceHolder))
                 .caseAdditionalSecurityFlag(HearingsCaseMapping.shouldBeAdditionalSecurityFlag(caseData))
                 .facilitiesRequired(HearingsDetailsMapping.getFacilitiesRequired())
                 .listingComments(HearingsDetailsMapping.getListingComments(caseData))
